+++ mobile/src/main/java/com/example/weibinwang/myapplication/Bean/VoteNote.java
@@ -5,7 +5,7 @@ package com.example.weibinwang.myapplication.Bean;
  */
 
 public class VoteNote {
-    private int id;
+    private long id;
     private User owner;
     private Restaurant purpose;
     private int note;
@@ -18,7 +18,7 @@ public class VoteNote {
     * The part of get() methods.
     * */
 
-    public int getId(){return this.id;}
+    public long getId(){return this.id;}
     public User owner(){return this.owner;}
     public Restaurant getPurpose(){return this.purpose;}
     public int getNote(){return this.note;}
@@ -26,7 +26,7 @@ public class VoteNote {
     /*
     * The part of set() methods.
     * */
-    public void setId(int id){
+    public void setId(long id){
         this.id = id;
     }
     public void setOwner(User owner){
